% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{build_country_list}
\alias{build_country_list}
\title{This function downloads the list of countries from the UN website.
The UN countries will be used as the basis for standardising country names in the NOAA file.
The function customises the UN list to better match the "countries" listed in the NOAA file.}
\usage{
build_country_list()
}
\value{
A data frame with a standardised list of countries and their equivalent in the NOAA file.
}
\description{
This function downloads the list of countries from the UN website.
The UN countries will be used as the basis for standardising country names in the NOAA file.
The function customises the UN list to better match the "countries" listed in the NOAA file.
}
\examples{
\dontrun{build_country_list()}
}
