% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFunctions.R
\name{eq_map}
\alias{eq_map}
\title{Plot earthquakes on a map.}
\usage{
eq_map(Data, annot_col = "Date")
}
\arguments{
\item{Data}{The data frame containing the earthquake data}

\item{annot_col}{The name of the column in Data to use to annotate the earthquakes on the map.}
}
\value{
A leaflet map showing the plotted earthquakes.
}
\description{
Plot earthquakes on a map.
}
\examples{
\dontrun{eq_map(Data=input_data, annot_col = "Location")}
}
