% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFunctions.R
\name{eq_map_label}
\alias{eq_map_label}
\title{Plot earthquakes on a map and includes custom labeling based on location, magnitude, and deaths.}
\usage{
eq_map_label(Data)
}
\arguments{
\item{Data}{The data frame containing the earthquake data}
}
\value{
A leaflet map showing the plotted earthquakes.
}
\description{
Plot earthquakes on a map and includes custom labeling based on location, magnitude, and deaths.
}
\examples{
\dontrun{eq_map(Data=input_data, annot_col = "Location")}
}
