% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotExtensions.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\alias{geom_timeline_label}
\title{GGplot function that creates a timeline graph of a group of earthquakes with associated labels}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomTimelineLabel

geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{ggplot aes parameter}

\item{data}{input data - dataframe of NOAA earthquakes}

\item{stat}{ggplot stat parameter}

\item{position}{ggplot position parameter}

\item{na.rm}{ggplot na.rm parameter}

\item{show.legend}{ggplot show.legend parameter}

\item{inherit.aes}{ggplot inherit.aes parameter}

\item{...}{ggplot other parameters}
}
\value{
Draws a GGplot graph
}
\description{
GGplot function that creates a timeline graph of a group of earthquakes with associated labels
}
\examples{
\dontrun{ggplot(inputData) + geom_timeline(aes(x=Date, colour=Deaths, label=Location, n_max=5, size=Mag))}
}
\keyword{datasets}
