% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{read_noaa_file}
\alias{read_noaa_file}
\title{This function reads a NOAA earthquake file.}
\usage{
read_noaa_file(input)
}
\arguments{
\item{input}{The location of a NOAA earthquake data file in tsv format.}
}
\value{
A data frame with a standardised list of NOAA earthquakes.
}
\description{
NOAA earthquake data can be downloaded from downloaded from https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/event-data.
}
\examples{
\dontrun{read_noaa_file(earthquakes)}
}
