% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotExtensions.R
\name{theme_quake}
\alias{theme_quake}
\title{GGplot function that creates a timeline graph of a group of earthquakes with associated labels}
\usage{
theme_quake()
}
\value{
Draws a GGplot graph
}
\description{
GGplot function that creates a timeline graph of a group of earthquakes with associated labels
}
\examples{
\dontrun{ggplot(inputData) + geom_timeline(aes(x=Date, colour=Deaths, label=Location, n_max=5, size=Mag)) + theme_quake()}
}
